refactor core module to make adding more methods easier to read
