+++ backend/src/core/__init__.py
@@ -3,16 +3,14 @@
 This file exports the core functions needed to run NSO Bridge.
 """
 
+from .database import CASCADE_CHILD, CASCADE_OTHER, BaseSQLModel, DatabaseEngine
 from .dependencies import AsyncSessionDepends, EngineFactory
-from .models import CASCADE_CHILD, CASCADE_OTHER, BaseSQLModel
+from .protocols import Memento
 from .router import api_router, assets, pages_router
-from .schemas import ClientSchema, ServerSchema
+from .schemas import APIResponseClass, ClientSchema, ServerSchema
 from .service import (
-    APIResponseClass,
-    DatabaseEngine,
-    Memento,
-    build_server,
     error_handlers,
+    get_server,
     shutdown,
 )
 from .utils import timedelta_serializer
@@ -23,13 +21,13 @@ __all__ = (
     'assets',
     'AsyncSessionDepends',
     'BaseSQLModel',
-    'build_server',
     'CASCADE_CHILD',
     'CASCADE_OTHER',
     'ClientSchema',
     'DatabaseEngine',
     'EngineFactory',
     'error_handlers',
+    'get_server',
     'Memento',
     'pages_router',
     'ServerSchema',
