+++ backend/src/core/service.py
@@ -2,198 +2,24 @@
 
 from __future__ import annotations
 
-import json
 import logging
 import os
 import signal
 from http import HTTPStatus
 from socket import AF_INET, SOCK_DGRAM, socket
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Callable,
-    ClassVar,
-    Final,
-    Mapping,
-    Protocol,
-    override,
-)
+from typing import TYPE_CHECKING, Any, Callable, Final
 
 from fastapi.exceptions import RequestValidationError
-from fastapi.responses import JSONResponse
-from sqlalchemy.engine import URL
-from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
-from sqlalchemy.orm import DeclarativeBase
 from uvicorn import Config, Server
 
 from .exceptions import ClientError, ModelLookupError
-from .schemas import APISchema, ErrorSchema
+from .schemas import APIResponseClass, ErrorSchema
 
 if TYPE_CHECKING:
     from fastapi import FastAPI, Request
-    from sqlalchemy.ext.asyncio import AsyncEngine
-    from sqlalchemy.orm import DeclarativeBase
-    from starlette.background import BackgroundTask
 
 
 logging.getLogger('aiosqlite').setLevel(logging.CRITICAL)
-
-SQLALCHEMY_DEBUG: bool = os.environ.get('SQLALCHEMY_DEBUG', '').lower() in {
-    'true',
-    'yes',
-}
-
-
-class Memento(Protocol):
-    """Represent a memento point-in-time of the application state.
-
-    Mementos can be used to implement functionality such as undo and redo by restoring
-    the application to a previous state.
-    """
-
-    async def restore(self) -> Memento:
-        """Restore the state of the application to when this Memento was constructed.
-
-        Returns:
-            Memento: A Memento of the state of the application before this method was
-            called. Calling `restore()` on this newly created Memento has the effect of
-            redoing an operation.
-
-        """
-        ...
-
-
-class DatabaseEngine:
-    """A connection to a database which stores models.
-
-    Attributes:
-        path (str): the relative path to the database.
-
-    """
-
-    _DRIVER: ClassVar[Final[str]] = 'sqlite+aiosqlite'
-
-    def __init__(self, db_schema: type[DeclarativeBase], db_path: str = '') -> None:
-        """Create a database engine without connecting to the database.
-
-        Args:
-            db_schema (type[DeclarativeBase]): a SQLAlchemy base model type which will
-            be initialized with the database.
-            db_path (str, optional): The relative path to the database. If left blank,
-            a database in memory will be used. Defaults to ''.
-
-        Raises:
-            ValueError: if the db_path is not a legal file name.
-
-        """
-        # TODO: ensure that path is a legal file name
-        if not db_path.isprintable():
-            logging.error('invalid database pathname')
-            raise ValueError('db path is invalid')
-        self._db_schema: type[DeclarativeBase] = db_schema
-        self._session_factory: async_sessionmaker[AsyncSession] | None = None
-        self.path: Final[str] = db_path if db_path != '' else ':memory:'
-
-    async def create_all(self) -> None:
-        """Initialize the connection to the database and create all tables.
-
-        Raises:
-            RuntimeError: if the database connection has already been established.
-
-        """
-        if self._session_factory is not None:
-            logging.error('the database has already been created')
-            raise RuntimeError('the database has already been created')
-
-        # Initialize the database engine
-        url: URL = URL.create(self._DRIVER, database=self.path)
-        logging.debug(f'Initializing engine at "{str(url)}"')
-        engine: AsyncEngine = create_async_engine(url, echo=SQLALCHEMY_DEBUG)
-
-        # Create the database tables
-        async with engine.connect() as session:
-            await session.run_sync(self._db_schema.metadata.create_all)
-        self._session_factory: async_sessionmaker[AsyncSession] = async_sessionmaker(
-            bind=engine, expire_on_commit=False
-        )
-
-    def is_connected(self) -> bool:
-        """Return True if the database is connected.
-
-        Returns:
-            bool: True if the database is connected.
-
-        """
-        return self._session_factory is not None
-
-    def get_async_session_factory(self) -> async_sessionmaker[AsyncSession]:
-        """Return a session factory that is associated with the database engine.
-
-        Raises:
-            RuntimeError: if the database has not yet been created.
-
-        Returns:
-            async_sessionmaker: an asynchronous session factory.
-
-        """
-        if self._session_factory is None:
-            raise RuntimeError('the database has not been created yet')
-        return self._session_factory
-
-    def get_async_session(self) -> AsyncSession:
-        """Return a session that is associated with the database engine.
-
-        Raises:
-            RuntimeError: if the database has not yet been created.
-
-        Returns:
-            AsyncSession: an asynchronous session.
-
-        """
-        factory: async_sessionmaker[AsyncSession] = self.get_async_session_factory()
-        return factory()
-
-
-class APIResponseClass(JSONResponse):
-    """Used to wrap all API responses in a common JSON interface.
-
-    See `core.schemas.APISchema`.
-    """
-
-    @override
-    def __init__(
-        self,
-        content: Any,
-        status_code: int = HTTPStatus.OK,
-        headers: Mapping[str, str] | None = None,
-        media_type: str | None = None,
-        background: BackgroundTask | None = None,
-    ) -> None:
-        error_occurred: bool = status_code not in range(
-            HTTPStatus.OK, HTTPStatus.MULTIPLE_CHOICES
-        )
-        super().__init__(
-            APISchema(
-                status_code=status_code,
-                error=content if error_occurred else None,
-                data=content if not error_occurred else None,
-            ).model_dump(),
-            status_code,
-            headers,
-            media_type,
-            background,
-        )
-
-    @override
-    def render(self, content: Any) -> bytes:
-        return json.dumps(
-            content,
-            ensure_ascii=False,
-            allow_nan=False,
-            indent=None,
-            separators=(',', ':'),
-            default=(lambda dt: str(dt)),  # Serialize datetime objects
-        ).encode('utf-8')
 
 
 async def _generic_error_handler(request: Request, e: Exception) -> APIResponseClass:
@@ -231,7 +57,7 @@ error_handlers: Final[dict[type[Exception], Callable[[Request, ...], Any]]] = {
 }
 
 
-def build_server(app: FastAPI) -> Server:
+def get_server(app: FastAPI) -> Server:
     """Build and return a Uvicorn server to serve the desired FastAPI app.
 
     Args:
